Created the upgrade section for the base
Created the upgrade section for the base, this included a new panel on the bases section which only appears during with bases available with a special ability.
The upgrades can be unlocked, and will show when in the game.

Currently there is no save section for bases and upgrades unlike players