Fixed issue with the base Template save
Base template type cause all of the base files to be re-named basetemplate causing a duplicated file issue, and wiping all of the object references on the main player choice scene. Should be working now, but the previous snapshot has a fully working version from before the changes if needed.